test(artifacts): remove skip if Nexus
