+++ tests/pytest_tests/system_tests/test_artifacts/test_public_api.py
@@ -5,7 +5,6 @@ import wandb
 from wandb import Api
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_fetching_artifact_files(user, wandb_init):
     project = "test"
 
@@ -37,7 +36,6 @@ def test_fetching_artifact_files(user, wandb_init):
     assert open(file_path).read() == "testing"
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_save_aliases_after_logging_artifact(user, wandb_init):
     project = "test"
     run = wandb_init(entity=user, project=project)
@@ -59,7 +57,6 @@ def test_save_aliases_after_logging_artifact(user, wandb_init):
     assert "hello" in aliases
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_update_aliases_on_artifact(user, wandb_init):
     project = "test"
     run = wandb_init(entity=user, project=project)
@@ -99,7 +96,6 @@ def test_update_aliases_on_artifact(user, wandb_init):
     assert "sequence" not in aliases
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_artifact_version(wandb_init):
     def create_test_artifact(content: str):
         art = wandb.Artifact("test-artifact", "test-type")
@@ -131,7 +127,6 @@ def test_artifact_version(wandb_init):
     assert artifact.source_version == "v1"
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_delete_collection(wandb_init):
     with wandb_init(project="test") as run:
         art = wandb.Artifact("test-artifact", "test-type")
@@ -167,7 +162,6 @@ def test_delete_collection(wandb_init):
         )
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_log_with_wrong_type_entity_project(wandb_init, logged_artifact):
     # todo: logged_artifact does not work with core
     entity, project = logged_artifact.entity, logged_artifact.project
@@ -191,7 +185,6 @@ def test_log_with_wrong_type_entity_project(wandb_init, logged_artifact):
             run.log_artifact(draft)
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_run_log_artifact(wandb_init):
     # Prepare data.
     with wandb_init() as run:
