+++ tests/pytest_tests/system_tests/test_core/test_keras_full.py
@@ -334,7 +334,6 @@ def test_keras_save_model(dummy_model, dummy_data, wandb_init):
     assert len(glob.glob(os.path.join(run.dir, "model-best.h5"))) == 1
 
 
-@pytest.mark.wandb_core_failure(feature="artifacts")
 def test_keras_dsviz(dummy_model, dummy_data, wandb_init):
     run = wandb_init()
     dummy_model.fit(
